% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_packages.R
\name{create_wp}
\alias{create_wp}
\title{Create Work Package directories}
\usage{
create_wp(wp_number)
}
\arguments{
\item{wp_number}{An integer representing the work package number.}
}
\value{
Nothing, called for its side effects.
}
\description{
This function creates the necessary subdirectories for a new work package (WP)
within the 'ANALYSES' and 'OUTPUT' folders.
}
\examples{
\dontrun{
create_wp(1)
}
}
